Consolidate the subheader setup into a subheader component
