+++ apps/timetable/admin/ui/js/index.js
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-define(['gh.core', 'gh.subheader', 'gh.calendar', 'gh.admin-listview', 'chosen', 'clickover', 'jquery-bbq'], function(gh) {
+define(['gh.core', 'gh.subheader', 'gh.calendar', 'gh.admin-listview', 'clickover', 'jquery-bbq'], function(gh) {
 
     var state = $.bbq.getState() || {};
 
@@ -42,10 +42,10 @@ define(['gh.core', 'gh.subheader', 'gh.calendar', 'gh.admin-listview', 'chosen',
             triposData = data;
 
             // Set up the tripos picker after all data has been retrieved
-            setUpTriposPicker();
-
-            // Run the hashchange logic to put the right selections in place
-            handleHashChange();
+            // Initialise the subheader component after all data has been retrieved
+            $(document).trigger('gh.subheader.init', {
+                'triposData': triposData
+            });
         });
     };
 
@@ -104,140 +104,6 @@ define(['gh.core', 'gh.subheader', 'gh.calendar', 'gh.admin-listview', 'chosen',
         gh.api.utilAPI.renderTemplate($('#gh-subheader-pickers-template'), {
             'gh': gh
         }, $('#gh-subheader'));
-
-        // Initialise the subheader component
-        $(document).trigger('gh.subheader.init');
-    };
-
-    /**
-     * Set up the modules of events in the sidebar
-     *
-     * @param  {Event}     ev      Standard jQuery event
-     * @param  {Object}    data    Data object describing the selected part to fetch modules for
-     * @private
-     */
-    var setUpModules = function(ev, data) {
-
-        // Hide the tripos help text
-        $('.gh-tripos-help').hide();
-        var partId = parseInt(data.selected, 10);
-
-        // Push the selected tripos in the URL
-        state['part'] = partId;
-        $.bbq.pushState(state);
-
-        // Track the part picker change in GA
-        gh.api.utilAPI.sendTrackingEvent('admin-picker', 'change', 'Admin part picker', partId);
-
-        gh.api.orgunitAPI.getOrgUnits(gh.data.me.AppId, true, partId, ['module'], function(err, modules) {
-            if (err) {
-                gh.api.utilAPI.notification('Fetching modules failed.', 'An error occurred while fetching the modules.', 'error');
-            }
-
-            // Sort the data before displaying it
-            modules.results.sort(gh.api.utilAPI.sortByDisplayName);
-            $.each(modules.results, function(i, module) {
-                module.Series.sort(gh.api.utilAPI.sortByDisplayName);
-            });
-
-            // Decorate the modules with their expanded status if LocalStorage is supported
-            var expandedIds = [];
-            if (Storage) {
-                expandedIds = _.compact(gh.api.utilAPI.localDataStorage().get('expanded'));
-                _.each(modules.results, function(module) {
-                    module.expanded = (_.indexOf(expandedIds, String(module.id)) > -1);
-                });
-            }
-
-            // Render the series in the sidebar
-            gh.api.utilAPI.renderTemplate($('#gh-modules-template'), {
-                'data': modules.results
-            }, $('#gh-modules-container'));
-
-            // Clear local storage
-            gh.api.utilAPI.localDataStorage().remove('expanded');
-
-            // Add the current expanded module(s) back to the local storage
-            expandedIds = _.compact([$('.gh-list-group-item-open').attr('data-id')]);
-            gh.api.utilAPI.localDataStorage().store('expanded', expandedIds);
-        });
-    };
-
-    /**
-     * Set up the part picker in the subheader
-     *
-     * @param  {Event}     ev      Standard jQuery event
-     * @param  {Object}    data    Data object describing the selected tripos to fetch parts for
-     * @private
-     */
-    var setUpPartPicker = function(ev, data) {
-        var triposId = parseInt(data.selected, 10);
-
-        // Push the selected tripos in the URL
-        state = {
-            'tripos': triposId,
-            'part': $.bbq.getState()['part']
-        };
-        $.bbq.pushState(state);
-
-        // Track the tripos picker change in GA
-        gh.api.utilAPI.sendTrackingEvent('admin-picker', 'change', 'Admin Tripos picker', triposId);
-
-        // Get the parts associated to the selected tripos
-        var parts = _.filter(triposData.parts, function(part) {
-            return parseInt(data.selected, 10) === part.ParentId;
-        });
-
-        // Render the results in the part picker
-        gh.api.utilAPI.renderTemplate($('#gh-subheader-part-template'), {
-            'data': parts
-        }, $('#gh-subheader-part'));
-
-        // Show the subheader part picker
-        $('#gh-subheader-part').show();
-
-        // Destroy the field if it's been initialised previously
-        $('#gh-subheader-part').chosen('destroy').off('change', setUpModules);
-
-        // Initialise the Chosen plugin on the part picker
-        $('#gh-subheader-part').chosen({
-            'no_results_text': 'No matches for',
-            'disable_search_threshold': 10
-        }).on('change', setUpModules);
-    };
-
-    /**
-     * Set up the Tripos picker in the subheader
-     *
-     * @private
-     */
-    var setUpTriposPicker = function() {
-        var triposPickerData = {
-            'courses': triposData.courses
-        };
-
-        _.each(triposPickerData.courses, function(course) {
-            course.subjects = _.filter(triposData.subjects, function(subject) {
-                return course.id === subject.ParentId;
-            });
-        });
-
-        // Massage the data so that courses are linked to their child subjects
-        // Render the results in the tripos picker
-        gh.api.utilAPI.renderTemplate($('#gh-subheader-picker-template'), {
-            'data': triposPickerData
-        }, $('#gh-subheader-tripos'));
-
-        // Show the subheader tripos picker
-        $('#gh-subheader-tripos').show();
-
-        // Destroy the field if it's been initialised previously
-        $('#gh-subheader-tripos').chosen('destroy').off('change', setUpModules);
-
-        // Initialise the Chosen plugin on the tripos picker
-        $('#gh-subheader-tripos').chosen({
-            'no_results_text': 'No matches for'
-        }).change(setUpPartPicker);
     };
 
     /**
@@ -297,44 +163,6 @@ define(['gh.core', 'gh.subheader', 'gh.calendar', 'gh.admin-listview', 'chosen',
                 gh.api.utilAPI.notification('Logout failed', 'Logging out of the application failed', 'error');
             }
         });
-    };
-
-    /**
-     * Handle the hashchange event by applying state values to the pickers. Can also be
-     * used separate from the hashchange event to apply state values to the pickers
-     *
-     * @private
-     */
-    var handleHashChange = function() {
-        state = $.bbq.getState() || {};
-
-        // If the URL shows a preselected tripos, select that tripos automatically
-        if (state.tripos && !_.isEmpty(state.tripos)) {
-            $('#gh-subheader-tripos').val(state.tripos);
-            $('#gh-subheader-tripos').trigger('change', {'selected': state.tripos});
-            $('#gh-subheader-tripos').trigger('chosen:updated');
-        } else {
-            // There is no state for the tripos, make sure it's reset
-            setUpTriposPicker();
-            // There can't be a part because there is no tripos
-            if ($('#gh_subheader_part_chosen').length) {
-                // Destroy the field if it's been initialised previously
-                $('#gh-subheader-part').chosen('destroy').off('change', setUpModules);
-                // Show the subheader part picker
-                $('#gh-subheader-part').hide();
-            }
-        }
-
-        // If the URL shows a preselected part, select that part automatically
-        if (state.part && $('#gh-subheader-part [value="' + state.part + '"]').length) {
-            $('#gh-subheader-part').val(state.part);
-            $('#gh-subheader-part').trigger('change', {'selected': state.part});
-            $('#gh-subheader-part').trigger('chosen:updated');
-        } else {
-            // Remove any modules and event series from the sidebar when no part is selected
-            // so no inaccurate information is presented to the user
-            $('#gh-modules-container').empty();
-        }
     };
 
 
@@ -408,9 +236,6 @@ define(['gh.core', 'gh.subheader', 'gh.calendar', 'gh.admin-listview', 'chosen',
         // Login and logout
         $('body').on('submit', '#gh-signin-form', doLogin);
         $('body').on('submit', '#gh-signout-form', doLogout);
-
-        // Handle hash changes
-        $(window).on('hashchange', handleHashChange);
     };
 
     /**
